Refactor(seokho/board): require 경로 적용 #493
